+++ CHANGELOG.rst
@@ -8,6 +8,8 @@ lib_mic_array change log
     to the mic_array so that they can benefit from the post processing of the
     mic_array.
   * Enabled the metadata which delivers the frame counter.
+  * Small fix to the filter generator to allow the use of fewer taps in the
+    final stage FIR.
   * Added significant bits collection to the metadata.
   * Added fixed gain control through define MIC_ARRAY_FIXED_GAIN.
   * Tested and enabled the debug mode for detecting frame dropping. Enabled by